Create a kitchsink test file for dumping tests
