+++ src/dom/__test__/kitchensink.test.js
@@ -1,11 +1,58 @@
+import $ from 'jquery'
 import {Observable} from 'rxjs/Observable'
-import {renderInDocument} from './support/renderInDocument'
-import {h} from '../h'
 import {BehaviorSubject} from 'rxjs/subject/BehaviorSubject'
+import {h} from '../h'
+import {renderInDocument} from './support/renderInDocument'
 
+import 'rxjs/add/operator/merge'
+import 'rxjs/add/operator/scan'
 import 'rxjs/add/operator/startWith'
 
-describe(`render`, () => {
+describe(`kitchen sink of tests`, () => {
+  it(`creates a simple clicker`, () => {
+    function Counter ({createEventHandler}) {
+      const handlePlus = createEventHandler(1)
+      const handleMinus = createEventHandler(-1)
+      const count = handlePlus
+                    .merge(handleMinus)
+                    .scan((x, y) => x + y)
+                    .startWith(0)
+
+      return h(`div`, {},
+        h(`button#plus`, {onClick: handlePlus}, `+ `, `PLUS`),
+        h(`button#minus`, {onClick: handleMinus}, `- `, `MINUS`),
+        h(`span`, {}, count),
+      )
+    }
+
+    const vnode = h(Counter)
+    const {node, cleanup} = renderInDocument(vnode)
+
+    const $node = $(node)
+    const $plus = $node.find(`#plus`)
+    const $minus = $node.find(`#minus`)
+
+    assert.equal(node.tagName, `DIV`)
+    assert.equal(node.children[0].tagName, `BUTTON`)
+    assert.equal(node.children[0].id, `plus`)
+    assert.equal(node.children[0].textContent, `+ PLUS`)
+    assert.equal(node.children[1].tagName, `BUTTON`)
+    assert.equal(node.children[1].id, `minus`)
+    assert.equal(node.children[1].textContent, `- MINUS`)
+    assert.equal(node.children[2].tagName, `SPAN`)
+    assert.equal(node.children[2].textContent, `0`)
+
+    $plus.trigger(`click`)
+    $plus.trigger(`click`)
+    $plus.trigger(`click`)
+    $plus.trigger(`click`)
+    $minus.trigger(`click`)
+
+    assert.equal(node.children[2].textContent, `3`)
+
+    cleanup()
+  })
+
   it(`renders a virtual node into a container`, () => {
     let mountCallbackCount = 0
     let unmountCallbackCount = 0
